Wrote unit tests for existing instructions

+++ src/cpu.h
@@ -8,7 +8,7 @@ typedef struct cpu_t {
     uint16_t pc;
     uint16_t previous_pc;
     uint16_t stack[16];
-    uint8_t sp;
+    int sp;
     uint8_t registers[16];
     uint16_t I;
     bool vf;
